Move the math subpackages into a single math package (re: #115)
