+++ packages/math/README.md
@@ -30,23 +30,9 @@ We're not able to support external contributors at this time, but check back in 
 
 ## Packages
 
-### math
-
-Module                | Description
---------------------- | -------------
-[@id-sdk/extent]      | 📦 Extent class for creating bounding boxes
-[@id-sdk/geo]         | 🌐 Geographic (spherical) math functions
-[@id-sdk/geom]        | 📈 Geometric (planar) math functions
-[@id-sdk/projection]  | 📽 Projection class for converting between Lon/Lat (λ,φ) and Cartesian (x,y) coordinates
-[@id-sdk/tiler]       | 🀄️ Tiler class for splitting the world into rectangular tiles
-[@id-sdk/vector]      | 📐 Vector (coordinate) math functions
----                   | ---
-[@id-sdk/math]        | 🔢 All of the above in one convenient package
-
-[@id-sdk/extent]: /packages/math/packages/extent
-[@id-sdk/geo]: /packages/math/packages/geo
-[@id-sdk/geom]: /packages/math/packages/geom
-[@id-sdk/projection]: /packages/math/packages/projection
-[@id-sdk/tiler]: /packages/math/packages/tiler
-[@id-sdk/vector]: /packages/math/packages/vector
-[@id-sdk/math]: /packages/math
+📦 Extent class for creating bounding boxes
+🌐 Geographic (spherical) math functions
+📈 Geometric (planar) math functions
+📽 Projection class for converting between Lon/Lat (λ,φ) and Cartesian (x,y) coordinates
+🀄️ Tiler class for splitting the world into rectangular tiles
+📐 Vector (coordinate) math functions
