+++ packages/math/esbuild.js
@@ -3,7 +3,7 @@ import { nodeExternalsPlugin } from 'esbuild-node-externals';
 
 esbuild.build({
   bundle: true,
-  entryPoints: ['./index.ts'],
+  entryPoints: ['./src/index.ts'],
   logLevel: 'info',
   outfile: './built/math.mjs',
   platform: 'neutral',
