--- packages/math/index.ts
@@ -1,6 +1,0 @@
-export * from '@id-sdk/extent';
-export * from '@id-sdk/geo';
-export * from '@id-sdk/geom';
-export * from '@id-sdk/projection';
-export * from '@id-sdk/tiler';
-export * from '@id-sdk/vector';
