--- packages/math/packages/extent/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "@id-sdk/extent",
-  "version": "3.0.0-pre.10",
-  "description": "Extent class for creating bounding boxes",
-  "homepage": "https://github.com/ideditor/id-sdk/tree/main/packages/math/extent#readme",
-  "repository": "github:ideditor/id-sdk",
-  "license": "ISC",
-  "contributors": [
-    "Ben Clark <clarkben@fb.com> (https://github.com/bonkles)",
-    "Bryan Housel <bhousel@gmail.com> (https://github.com/bhousel)",
-    "Danny Krause (https://github.com/vanreece)",
-    "Milos Brzakovic <mbrzakovic@microsoft.com> (https://github.com/mbrzakovic)"
-  ],
-  "files": [
-    "built/",
-    "src/"
-  ],
-  "type": "module",
-  "typedocMain" : "./src/extent.ts",
-  "source": "./src/extent.ts",
-  "types": "./built/extent.d.ts",
-  "main": "./built/extent.mjs",
-  "module": "./built/extent.mjs",
-  "scripts": {
-    "clean": "shx rm -rf ./built",
-    "build": "run-p esbuild tsc",
-    "esbuild": "node ./esbuild.js",
-    "tsc": "tsc",
-    "test": "NODE_OPTIONS=--experimental-vm-modules npx jest --colors --config=../../../../package.json --roots=./packages/math/packages/extent"
-  },
-  "dependencies": {
-    "@id-sdk/geo": "^3.0.0-pre.10",
-    "@id-sdk/vector": "^3.0.0-pre.10"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
