--- packages/math/packages/geo/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@id-sdk/geo",
-  "version": "3.0.0-pre.10",
-  "description": "Geographic (spherical) math functions",
-  "homepage": "https://github.com/ideditor/id-sdk/tree/main/packages/math/geo#readme",
-  "repository": "github:ideditor/id-sdk",
-  "license": "ISC",
-  "contributors": [
-    "Ben Clark <clarkben@fb.com> (https://github.com/bonkles)",
-    "Bryan Housel <bhousel@gmail.com> (https://github.com/bhousel)",
-    "Danny Krause (https://github.com/vanreece)",
-    "Milos Brzakovic <mbrzakovic@microsoft.com> (https://github.com/mbrzakovic)"
-  ],
-  "files": [
-    "built/",
-    "src/"
-  ],
-  "type": "module",
-  "typedocMain" : "./src/geo.ts", 
-  "source": "./src/geo.ts",
-  "types": "./built/geo.d.ts",
-  "main": "./built/geo.mjs",
-  "module": "./built/geo.mjs",
-  "scripts": {
-    "clean": "shx rm -rf ./built",
-    "build": "run-p esbuild tsc",
-    "esbuild": "node ./esbuild.js",
-    "tsc": "tsc",
-    "test": "NODE_OPTIONS=--experimental-vm-modules npx jest --colors --config=../../../../package.json --roots=./packages/math/packages/geo"
-  },
-  "dependencies": {
-    "@id-sdk/vector": "^3.0.0-pre.10"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
