--- packages/math/packages/geom/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@id-sdk/geom",
-  "version": "3.0.0-pre.10",
-  "description": "Geometric (planar) math functions",
-  "homepage": "https://github.com/ideditor/id-sdk/tree/main/packages/math/geom#readme",
-  "repository": "github:ideditor/id-sdk",
-  "license": "ISC",
-  "contributors": [
-    "Ben Clark <clarkben@fb.com> (https://github.com/bonkles)",
-    "Bryan Housel <bhousel@gmail.com> (https://github.com/bhousel)",
-    "Danny Krause (https://github.com/vanreece)",
-    "Milos Brzakovic <mbrzakovic@microsoft.com> (https://github.com/mbrzakovic)"
-  ],
-  "files": [
-    "built/",
-    "src/"
-  ],
-  "type": "module",
-  "typedocMain" : "./src/geom.ts",
-  "source": "./src/geom.ts",
-  "types": "./built/geom.d.ts",
-  "main": "./built/geom.mjs",
-  "module": "./built/geom.mjs",
-  "scripts": {
-    "clean": "shx rm -rf ./built",
-    "build": "run-p esbuild tsc",
-    "esbuild": "node ./esbuild.js",
-    "tsc": "tsc",
-    "test": "NODE_OPTIONS=--experimental-vm-modules npx jest --colors --config=../../../../package.json --roots=./packages/math/packages/geom"
-  },
-  "dependencies": {
-    "@id-sdk/extent": "^3.0.0-pre.10",
-    "@id-sdk/vector": "^3.0.0-pre.10",
-    "@types/d3-polygon": "^3.0.0",
-    "d3-polygon": "^3.0.1"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
