--- packages/math/packages/geom/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../../../tsconfig.json",
-  "include": ["src/"],
-  "exclude": ["node_modules/", "tests/"],
-  "compilerOptions": {
-    "outDir": "./built"
-  }
-}
