--- packages/math/packages/projection/esbuild.js
@@ -1,11 +1,0 @@
-import esbuild from 'esbuild';
-import { nodeExternalsPlugin } from 'esbuild-node-externals';
-
-esbuild.build({
-  bundle: true,
-  entryPoints: ['./src/projection.ts'],
-  logLevel: 'info',
-  outfile: './built/projection.mjs',
-  platform: 'neutral',
-  plugins: [nodeExternalsPlugin()]
-}).catch(() => process.exit(1))
