--- packages/math/packages/projection/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "@id-sdk/projection",
-  "version": "3.0.0-pre.10",
-  "description": "Projection class for converting between Lon/Lat (λ,φ) and Cartesian (x,y) coordinates",
-  "homepage": "https://github.com/ideditor/id-sdk/tree/main/packages/math/projection#readme",
-  "repository": "github:ideditor/id-sdk",
-  "license": "ISC",
-  "contributors": [
-    "Ben Clark <clarkben@fb.com> (https://github.com/bonkles)",
-    "Bryan Housel <bhousel@gmail.com> (https://github.com/bhousel)",
-    "Danny Krause (https://github.com/vanreece)",
-    "Milos Brzakovic <mbrzakovic@microsoft.com> (https://github.com/mbrzakovic)"
-  ],
-  "files": [
-    "built/",
-    "src/"
-  ],
-  "type": "module",
-  "typedocMain" : "./src/projection.ts",
-  "source": "./src/projection.ts",
-  "types": "./built/projection.d.ts",
-  "main": "./built/projection.mjs",
-  "module": "./built/projection.mjs",
-  "scripts": {
-    "clean": "shx rm -rf ./built",
-    "build": "run-p esbuild tsc",
-    "esbuild": "node ./esbuild.js",
-    "tsc": "tsc",
-    "test": "NODE_OPTIONS=--experimental-vm-modules npx jest --colors --config=../../../../package.json --roots=./packages/math/packages/projection"
-  },
-  "dependencies": {
-    "@id-sdk/vector": "^3.0.0-pre.10",
-    "@types/d3-geo": "^3.0.2",
-    "@types/d3-zoom": "^3.0.1",
-    "d3-geo": "^3.0.1",
-    "d3-zoom": "^3.0.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
