--- packages/math/packages/tiler/README.md
@@ -1,19 +1,0 @@
-[![npm version](https://badge.fury.io/js/%40id-sdk%2Ftiler.svg)](https://badge.fury.io/js/%40id-sdk%2Ftiler)
-
-## Installing
-
-`npm install @id-sdk/tiler`
-
-This library is distributed in ESM format only.  It cannot be `require()`'d from CommonJS.
-For more, please read Sindre Sorhus’s [FAQ](https://gist.github.com/sindresorhus/a39789f98801d908bbc7ff3ecc99d99c).
-
-```js
-import { Tiler } from '@id-sdk/tiler';
-```
-
-
-## Contributing
-
-This project is just getting started! 🌱
-
-We're not able to support external contributors at this time, but check back in a bit when things have matured.
