--- packages/math/packages/tiler/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@id-sdk/tiler",
-  "version": "3.0.0-pre.10",
-  "description": "Tiler class for splitting the world into rectangular tiles",
-  "homepage": "https://github.com/ideditor/id-sdk/tree/main/packages/math/tiler#readme",
-  "repository": "github:ideditor/id-sdk",
-  "license": "ISC",
-  "contributors": [
-    "Ben Clark <clarkben@fb.com> (https://github.com/bonkles)",
-    "Bryan Housel <bhousel@gmail.com> (https://github.com/bhousel)",
-    "Danny Krause (https://github.com/vanreece)",
-    "Milos Brzakovic <mbrzakovic@microsoft.com> (https://github.com/mbrzakovic)"
-  ],
-  "files": [
-    "built/",
-    "src/"
-  ],
-  "type": "module",
-  "typedocMain" : "./src/tiler.ts", 
-  "source": "./src/tiler.ts",
-  "types": "./built/tiler.d.ts",
-  "main": "./built/tiler.mjs",
-  "module": "./built/tiler.mjs",
-  "scripts": {
-    "clean": "shx rm -rf ./built",
-    "build": "run-p esbuild tsc",
-    "esbuild": "node ./esbuild.js",
-    "tsc": "tsc",
-    "test": "NODE_OPTIONS=--experimental-vm-modules npx jest --colors --config=../../../../package.json --roots=./packages/math/packages/tiler"
-  },
-  "dependencies": {
-    "@id-sdk/extent": "^3.0.0-pre.10",
-    "@id-sdk/geo": "^3.0.0-pre.10",
-    "@id-sdk/projection": "^3.0.0-pre.10",
-    "@id-sdk/vector": "^3.0.0-pre.10"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
