--- packages/math/packages/vector/README.md
@@ -1,20 +1,0 @@
-[![npm version](https://badge.fury.io/js/%40id-sdk%2Fvector.svg)](https://badge.fury.io/js/%40id-sdk%2Fvector)
-
-## Installing
-
-`npm install @id-sdk/vector`
-
-This library is distributed in ESM format only.  It cannot be `require()`'d from CommonJS.
-For more, please read Sindre Sorhus’s [FAQ](https://gist.github.com/sindresorhus/a39789f98801d908bbc7ff3ecc99d99c).
-
-```js
-import * as vec from '@id-sdk/vector';
-import { vecEqual } from '@id-sdk/vector';
-```
-
-
-## Contributing
-
-This project is just getting started! 🌱
-
-We're not able to support external contributors at this time, but check back in a bit when things have matured.
