--- packages/math/packages/vector/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@id-sdk/vector",
-  "version": "3.0.0-pre.10",
-  "description": "Vector (coordinate) math functions",
-  "homepage": "https://github.com/ideditor/id-sdk/tree/main/packages/math/vector#readme",
-  "repository": "github:ideditor/id-sdk",
-  "license": "ISC",
-  "contributors": [
-    "Ben Clark <clarkben@fb.com> (https://github.com/bonkles)",
-    "Bryan Housel <bhousel@gmail.com> (https://github.com/bhousel)",
-    "Danny Krause (https://github.com/vanreece)",
-    "Milos Brzakovic <mbrzakovic@microsoft.com> (https://github.com/mbrzakovic)"
-  ],
-  "files": [
-    "built/",
-    "src/"
-  ],
-  "type": "module",
-  "typedocMain" : "./src/vector.ts",
-  "source": "./src/vector.ts",
-  "types": "./built/vector.d.ts",
-  "main": "./built/vector.mjs",
-  "module": "./built/vector.mjs",
-  "scripts": {
-    "clean": "shx rm -rf ./built",
-    "build": "run-p esbuild tsc",
-    "esbuild": "node ./esbuild.js",
-    "tsc": "tsc",
-    "test": "NODE_OPTIONS=--experimental-vm-modules npx jest --colors --config=../../../../package.json --roots=./packages/math/packages/vector"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
