+++ packages/math/src/index.ts
@@ -1,0 +1,11 @@
+/**
+ * 🧰 Collection of iD-sdk math libraries
+ * @module
+ */
+
+export * from './extent';
+export * from './geo';
+export * from './geom';
+export * from './projection';
+export * from './tiler';
+export * from './vector';
