+++ packages/math/src/projection.ts
@@ -9,7 +9,7 @@
 
 import { geoMercatorRaw as d3_geoMercatorRaw, geoTransform as d3_geoTransform } from 'd3-geo';
 import { zoomIdentity as d3_zoomIdentity } from 'd3-zoom';
-import { Vec2 } from '@id-sdk/vector';
+import { Vec2 } from './vector';
 
 /** An Object containing `x`, `y`, `k` numbers */
 export interface Transform {
