+++ packages/math/src/vector.ts
