+++ packages/math/tests/extent.test.ts
@@ -1,6 +1,5 @@
-import 'jest-extended';
 import { Extent } from '../src/extent';
-import { Vec2 } from '@id-sdk/vector';
+import { Vec2 } from '../src/vector';
 
 describe('math/extent', () => {
   describe('constructor', () => {
