+++ packages/math/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["index.ts"],
+  "include": ["src/"],
   "exclude": ["node_modules/", "tests/"],
   "compilerOptions": {
     "outDir": "./built"
