+++ packages/util/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@id-sdk/util",
   "version": "3.0.0-pre.10",
-  "description": "util",
+  "description": "Collection of iD-sdk utility libraries",
   "homepage": "https://github.com/ideditor/id-sdk/tree/main/packages/util#readme",
   "repository": "github:ideditor/id-sdk",
   "license": "ISC",
@@ -17,7 +17,7 @@
   ],
   "type": "module",
   "typedocMain" : "./src/index.ts",
-  "source": "./src/index.js",
+  "source": "./src/index.ts",
   "types": "./built/index.d.ts",
   "main": "./built/util.mjs",
   "module": "./built/util.mjs",
