+++ packages/util/tests/osm.test.ts
@@ -1,4 +1,3 @@
-import 'jest-extended';
 import * as util from '../src/index';
 
 describe('utilCleanTags', () => {
